# src/services/api-gateway/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VEDO Core REST API",
    "description": "REST API for ontology management — classes, properties, individuals, versioning, and queries.\n\nBase URL: `/api/v1`\n\nAll endpoints require authentication via Bearer JWT token in the `Authorization` header.\n\n## Error format\n\nAll errors return:\n```json\n{\n  \"error\": {\n    \"code\": \"GATEWAY-ERROR-CODE\",\n    \"message\": \"Human-readable error description\"\n  }\n}\n```\n\n## Pagination\n\nList endpoints accept `page` and `per_page` query parameters and return:\n```json\n{\n  \"data\": [...],\n  \"meta\": {\n    \"page\": 1,\n    \"per_page\": 20,\n    \"total\": 42,\n    \"total_pages\": 3\n  }\n}\n```",
    "version": "0.2.0",
    "contact": {
      "name": "VEDO Core Team",
      "url": "https://vedo.dev"
    }
  },
  "servers": [
    {
      "url": "https://api.vedo.dev/api/v1",
      "description": "Production API"
    },
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Local development"
    }
  ],
  "paths": {
    "/ontologies": {
      "get": {
        "summary": "List ontologies",
        "description": "Returns a paginated list of ontologies the authenticated user has access to.",
        "operationId": "listOntologies",
        "tags": ["Ontologies"],
        "parameters": [
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/PerPage" },
          {
            "name": "q",
            "in": "query",
            "description": "Full-text search query",
            "schema": { "type": "string" },
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of ontologies",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/OntologySummary" }
                    },
                    "meta": { "$ref": "#/components/schemas/PaginationMeta" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/ontologies/{id}": {
      "get": {
        "summary": "Get ontology details",
        "description": "Returns full ontology details including counts of classes, properties, and individuals.",
        "operationId": "getOntology",
        "tags": ["Ontologies"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Ontology ID",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Ontology details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "$ref": "#/components/schemas/OntologyDetail" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Ontology not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/ontologies/{id}/classes": {
      "get": {
        "summary": "List classes",
        "description": "Returns a paginated list of classes in an ontology. Supports text search and parent filter.",
        "operationId": "listClasses",
        "tags": ["Classes"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Ontology ID",
            "required": true,
            "schema": { "type": "string" }
          },
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/PerPage" },
          {
            "name": "q",
            "in": "query",
            "description": "Full-text search query against class labels and comments",
            "schema": { "type": "string" },
            "required": false
          },
          {
            "name": "parent_id",
            "in": "query",
            "description": "Filter by parent class ID (returns direct children)",
            "schema": { "type": "string" },
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of classes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/ClassSummary" }
                    },
                    "meta": { "$ref": "#/components/schemas/PaginationMeta" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Ontology not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/ontologies/{id}/classes/{classId}": {
      "get": {
        "summary": "Get class details",
        "description": "Returns full class details including hierarchy breadcrumb, property count, and individual count.",
        "operationId": "getClass",
        "tags": ["Classes"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Ontology ID",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "classId",
            "in": "path",
            "description": "Class ID",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Class details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "$ref": "#/components/schemas/ClassDetail" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Ontology or class not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/ontologies/{id}/properties": {
      "get": {
        "summary": "List properties",
        "description": "Returns a paginated list of properties. Can be filtered by type (object or datatype).",
        "operationId": "listProperties",
        "tags": ["Properties"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Ontology ID",
            "required": true,
            "schema": { "type": "string" }
          },
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/PerPage" },
          {
            "name": "type",
            "in": "query",
            "description": "Filter by property type",
            "schema": {
              "type": "string",
              "enum": ["object", "datatype"]
            },
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of properties",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/PropertySummary" }
                    },
                    "meta": { "$ref": "#/components/schemas/PaginationMeta" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Ontology not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/ontologies/{id}/individuals": {
      "get": {
        "summary": "List individuals",
        "description": "Returns a paginated list of individuals. Can be filtered by class.",
        "operationId": "listIndividuals",
        "tags": ["Individuals"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Ontology ID",
            "required": true,
            "schema": { "type": "string" }
          },
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/PerPage" },
          {
            "name": "class_id",
            "in": "query",
            "description": "Filter by class ID (returns individuals of that class)",
            "schema": { "type": "string" },
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of individuals",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/IndividualSummary" }
                    },
                    "meta": { "$ref": "#/components/schemas/PaginationMeta" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Ontology not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/sparql": {
      "post": {
        "summary": "Execute SPARQL query",
        "description": "Execute a read-only SPARQL query against an ontology. Mutation queries (INSERT, DELETE) are rejected.\n\nA mandatory LIMIT of 1000 is enforced if no LIMIT clause is present.",
        "operationId": "executeSPARQL",
        "tags": ["Queries"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "query": {
                    "type": "string",
                    "description": "SPARQL 1.1 SELECT query"
                  }
                },
                "required": ["query"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Query results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "results": { "type": "object" },
                    "execution_time_ms": { "type": "integer" },
                    "triple_count": { "type": "integer" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid or mutation query",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/cypher": {
      "post": {
        "summary": "Execute CYPHER query",
        "description": "Execute a read-only CYPHER query against an ontology. Mutation queries are rejected.\n\nA mandatory LIMIT of 1000 is enforced if no LIMIT clause is present.",
        "operationId": "executeCYPHER",
        "tags": ["Queries"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "query": {
                    "type": "string",
                    "description": "CYPHER READ query"
                  }
                },
                "required": ["query"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Query results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "results": { "type": "object" },
                    "execution_time_ms": { "type": "integer" },
                    "triple_count": { "type": "integer" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid or mutation query",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/openapi.json": {
      "get": {
        "summary": "OpenAPI specification",
        "description": "Returns the OpenAPI 3.1 specification for this API.",
        "operationId": "getOpenAPISpec",
        "tags": ["Meta"],
        "responses": {
          "200": {
            "description": "OpenAPI specification",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "The complete OpenAPI 3.1 specification document"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "parameters": {
      "Page": {
        "name": "page",
        "in": "query",
        "description": "Page number (1-based)",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "default": 1
        },
        "required": false
      },
      "PerPage": {
        "name": "per_page",
        "in": "query",
        "description": "Results per page (max 100)",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 20
        },
        "required": false
      }
    },
    "schemas": {
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "page": { "type": "integer", "description": "Current page number" },
          "per_page": { "type": "integer", "description": "Items per page" },
          "total": { "type": "integer", "description": "Total number of items" },
          "total_pages": { "type": "integer", "description": "Total number of pages" }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": { "type": "string", "description": "Domain-prefixed error code" },
              "message": { "type": "string", "description": "Human-readable error message" }
            }
          }
        }
      },
      "OntologySummary": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "label": { "type": "string" },
          "description": { "type": "string" },
          "class_count": { "type": "integer" },
          "property_count": { "type": "integer" },
          "individual_count": { "type": "integer" },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
      "OntologyDetail": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "label": { "type": "string" },
          "description": { "type": "string" },
          "version": { "type": "string" },
          "class_count": { "type": "integer" },
          "property_count": { "type": "integer" },
          "individual_count": { "type": "integer" },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
      "ClassSummary": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "label": { "type": "string" },
          "comment": { "type": "string" },
          "parent_id": { "type": "string" },
          "child_count": { "type": "integer" }
        }
      },
      "ClassDetail": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "label": { "type": "string" },
          "comment": { "type": "string" },
          "iri": { "type": "string" },
          "parent_ids": { "type": "array", "items": { "type": "string" } },
          "breadcrumb": { "type": "array", "items": { "type": "string" } },
          "property_count": { "type": "integer" },
          "individual_count": { "type": "integer" }
        }
      },
      "PropertySummary": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "label": { "type": "string" },
          "type": { "type": "string", "enum": ["object", "datatype"] },
          "domain": { "type": "string" },
          "range": { "type": "string" },
          "is_functional": { "type": "boolean" }
        }
      },
      "IndividualSummary": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "label": { "type": "string" },
          "class_id": { "type": "string" },
          "class_name": { "type": "string" },
          "property_count": { "type": "integer" }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Keycloak-issued JWT token"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}